last checkpoint: include cities in the nested set model calculation
Previously, the contents of the country files were read and inserted into the database in chunks of 1000.
Then the Nested Set Model is built by reading the hierarchy.txt file and updating the records in the database(slow).

At this point, the Nested Set Model is built first by reading the hierarchy.txt file, and combining it with the contents
of the admin2Codes.txt file which contains the hierarchy for cities(Even though geonames.org has up to ADM5 division,
it cuts-off at ADM2 for cities). After the nestedSet is built, the country files contents is mapped into a Geoname with
the nestedSet properties(_lft, _rgt, depth and parent_id) and inserted into the database